refactor(tape): close storage bypasses
