+++ src/main/data/schemaCatalog.ts
@@ -19,9 +19,9 @@ import { DeepChatMessageSearchResultsTable } from '@/session/data/tables/deepcha
 import { DeepChatSearchDocumentsTable } from '@/session/data/tables/deepchatSearchDocuments'
 import { DeepChatPendingInputsTable } from '@/session/data/tables/deepchatPendingInputs'
 import { DeepChatUsageStatsTable } from '@/session/data/tables/deepchatUsageStats'
-import { DeepChatTapeEntriesTable } from '@/session/data/tables/deepchatTapeEntries'
+import { DeepChatTapeEntriesTable } from '@/tape/infrastructure/sqlite/tapeEntryStore'
 import { DeepChatMemoryIngestionProjectionTable } from '@/memory/data/tables/deepchatMemoryIngestionProjection'
-import { DeepChatTapeSearchProjectionTable } from '@/session/data/tables/deepchatTapeSearchProjection'
+import { DeepChatTapeSearchProjectionTable } from '@/tape/infrastructure/sqlite/tapeSearchProjectionStore'
 import { DeepChatSessionMetadataTable } from '@/session/data/tables/deepchatSessionMetadata'
 import { LegacyImportStatusTable } from '@/app/data/tables/legacyImportStatus'
 import { AgentsTable } from '@/agent/data/tables/agents'
